Add VenvProvisioner seam and move engine-generic utils out of workflows/
Introduce workflow_module.venv_provisioner with a VenvProvisioner
protocol (has_venv / venv_path / venv_python / provision) and a
process-wide registry with a lazy Tenstorrent default marked as the
Phase-2 removal point. The Tenstorrent implementation adapts
VENV_CONFIGS, preserving VenvConfig.setup(model_spec=...) semantics.

WorkflowVenvType moves to workflow_module/engine_types.py (the keys are
engine-generic; the venv *content* stays adapter-side) and is
re-exported from workflows/workflow_types.py. Engine-generic helpers
leave workflows/utils.py: run_command/stream_subprocess_output move to
workflow_module/proc.py, and the run-context knob helpers
(get_num_calls, is_sdxl_num_prompts_enabled, the whisper toggles, SDXL
constants) move to workflow_module/context_helpers.py; workflows/utils.py
re-exports both groups for pre-extraction callers.

All engine call sites (VenvCommand, command_factory release venvs,
eval_command, agentic driver, whisper/embedding/media tests) now go
through the provisioner or the new engine modules.
