+++ test_module/benchmark_tests/tts_benchmark_tests.py
@@ -19,7 +19,7 @@ _PROJECT_ROOT = Path(__file__).resolve().parents[2]
 if str(_PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(_PROJECT_ROOT))
 
-from workflows.utils import get_num_calls
+from workflow_module.context_helpers import get_num_calls
 
 from report_module.schema import Block
