+++ test_module/llm_tests/llm_eval_tests.py
@@ -22,7 +22,7 @@ from reference_config.evals.eval_config import accept_eval_score, resolve_eval_r
 from report_module.schema import Block
 from workflow_module import accept_blocks
 from workflow_module.engine_types import EvalLimitMode
-from workflows.utils import run_command
+from workflow_module.proc import run_command
 
 from .._test_common import ReportCheckTypes, TestStatus, block_id
 from ..context import MediaContext
