+++ workflows/utils.py
@@ -8,21 +8,25 @@ import base64
 import logging
 import os
 import re
-import shlex
-import subprocess
 import tempfile
-import threading
 import uuid
 from pathlib import Path
 from typing import Dict, List, Optional, Tuple
 
 logger = logging.getLogger(__name__)
 
-# SDXL num prompts limits
-SDXL_DEFAULT_NUM_PROMPTS = 100
-SDXL_LOWER_BOUND_NUM_PROMPTS = 2
-SDXL_UPPER_BOUND_NUM_PROMPTS = 5000
-
+# Engine-generic helpers live in the engine core; re-exported here for
+# pre-extraction callers (v1 workflows and tests).
+from workflow_module.context_helpers import (  # noqa: E402,F401
+    SDXL_DEFAULT_NUM_PROMPTS,
+    SDXL_LOWER_BOUND_NUM_PROMPTS,
+    SDXL_UPPER_BOUND_NUM_PROMPTS,
+    get_num_calls,
+    is_preprocessing_enabled_for_whisper,
+    is_sdxl_num_prompts_enabled,
+    is_streaming_enabled_for_whisper,
+)
+from workflow_module.proc import run_command, stream_subprocess_output  # noqa: E402,F401
 
 # Minimum image version supported by this run.py. The vLLM docker image
 # interface was reshaped in v0.11.0 (commit 50db8ac7 "Simplify and improve
@@ -217,120 +221,6 @@ def ensure_readwriteable_dir(path, raise_on_fail=True, logger=logger):
         return False
 
 
-def stream_subprocess_output(pipe, logger, level):
-    with pipe:
-        for line in iter(pipe.readline, ""):
-            logger.log(level, line.strip(), extra={"raw": True})
-
-
-def run_command(
-    command,
-    logger,
-    log_file_path=None,
-    shell=False,
-    copy_env=True,
-    env=None,
-    check=False,
-):
-    """
-    This function is a wrapper around subprocess.Popen and subprocess.run.
-    It is used to run a command and capture the stdout and stderr in the caller's logger.
-
-    Args:
-        command: Command to run. Can be a string or a list of strings.
-        logger: Logger to use for logging. Must be passed because the common use case is to capture the command's stdout and stderr in the caller's logger.
-        log_file_path: Path to log file. If None, stdout and stderr will be logged to the logger.
-        shell: Whether to use shell to run the command.
-        copy_env: Whether to copy the environment variables.
-        env: Environment variables to use.
-        check: Whether to check the return code. Set to True for commands that must succeed.
-    Returns:
-        Return code of the command.
-    Raises:
-        RuntimeError: If the command fails and check is True.
-        NotImplementedError: If copy_env is True and not implemented.
-        AssertionError: If command is not a list of strings.
-        ValueError: If command is None.
-        PermissionError: If the directory is not writable.
-        IOError: If the directory is not readable.
-    """
-    if not copy_env:
-        raise NotImplementedError("TODO")
-
-    if not env:
-        env = os.environ.copy()
-    # TODO: force usage to always use argument list
-    # use shlex to log full command before running
-
-    if command is None:
-        logger.error("No command provided to run_command.")
-    elif isinstance(command, str):
-        command = shlex.split(command)
-
-    assert isinstance(command, list), "Command must be a list of cmd arguments."
-
-    logger.info(f"Running command: {shlex.join(command)}")
-
-    if not log_file_path:
-        subproc_type = "subprocess.Popen"
-        # capture all output to stdout and stderr in current process
-        process = subprocess.Popen(
-            command,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            bufsize=1,
-            text=True,
-            env=env,
-        )
-
-        stdout_thread = threading.Thread(
-            target=stream_subprocess_output,
-            args=(process.stdout, logger, logging.DEBUG),
-        )
-        stderr_thread = threading.Thread(
-            target=stream_subprocess_output,
-            args=(process.stderr, logger, logging.INFO),
-        )
-
-        stdout_thread.start()
-        stderr_thread.start()
-
-        stdout_thread.join()
-        stderr_thread.join()
-
-        process.wait()
-        return_code = process.returncode
-    else:
-        subproc_type = "subprocess.run"
-        logger.info(f"Logging output to: {log_file_path} ...")
-        with open(log_file_path, "a", buffering=1) as log_file:
-            result = subprocess.run(
-                command,
-                shell=shell,
-                stdout=log_file,
-                stderr=log_file,
-                check=check,
-                text=True,
-                env=env,
-            )
-            return_code = result.returncode
-
-    if return_code != 0:
-        error_message = (
-            f"⛔ {subproc_type} command failed with return code: {return_code}\n"
-            f"command: {shlex.join(command)}\n\n"
-            "See error messages in logs above this RuntimeError for details on actual cause of failure.\n"
-        )
-        if check:
-            raise RuntimeError(error_message)
-        else:
-            logger.error(
-                error_message
-                + "\nThis command is optional or can be recovered from failure (check=False set). Continuing ...\n"
-            )
-    return return_code
-
-
 default_dotenv_path = get_repo_root_path() / ".env"
 
 
@@ -449,81 +339,6 @@ def resolve_hf_snapshot_dir(hf_repo: str, hf_home: Path) -> Optional[Path]:
 def get_weights_hf_cache_dir(hf_repo: str) -> Path:
     hf_home = get_default_hf_home_path()
     return resolve_hf_snapshot_dir(hf_repo, hf_home)
-
-
-def is_streaming_enabled_for_whisper(self) -> bool:
-    """Determine if streaming is enabled for the Whisper model based on CLI args. Default to True if not set."""
-    logger.info("Checking if streaming is enabled for Whisper model")
-    cli_args = getattr(self.model_spec, "cli_args", {})
-
-    # Check if streaming arg exists and has a valid value
-    streaming_value = cli_args.get("streaming")
-    if streaming_value is None:
-        return True
-
-    # Convert to string and check if it's 'true'
-    streaming_enabled = str(streaming_value).lower() == "true"
-
-    return streaming_enabled
-
-
-def is_preprocessing_enabled_for_whisper(self) -> bool:
-    """Determine if preprocessing is enabled for the Whisper model based on CLI args. Default to True if not set."""
-    logger.info("Checking if preprocessing is enabled for Whisper model")
-
-    cli_args = getattr(self.model_spec, "cli_args", {})
-    preprocessing_value = cli_args.get("preprocessing")
-    if preprocessing_value is None:
-        return True
-
-    # Convert to string and check if it's 'true'
-    preprocessing_enabled = str(preprocessing_value).lower() == "true"
-
-    return preprocessing_enabled
-
-
-def is_sdxl_num_prompts_enabled(self) -> int:
-    """Determine the number of prompts to use for SDXL based on CLI args. Default to 100 if not set."""
-    logger.info("Checking if sdxl_num_prompts is set")
-
-    cli_args = getattr(self.model_spec, "cli_args", {})
-    sdxl_num_prompts = cli_args.get("sdxl_num_prompts")
-    if sdxl_num_prompts is None:
-        return SDXL_DEFAULT_NUM_PROMPTS
-
-    # Convert to int and return
-    num_prompts = int(sdxl_num_prompts)
-    if (
-        num_prompts < SDXL_LOWER_BOUND_NUM_PROMPTS
-        or num_prompts > SDXL_UPPER_BOUND_NUM_PROMPTS
-    ):
-        return SDXL_DEFAULT_NUM_PROMPTS
-
-    return num_prompts
-
-
-def get_num_calls(self) -> int:
-    """Get number of calls from benchmark parameters."""
-    logger.info("Extracting number of calls from benchmark parameters")
-
-    # Guard clause: Handle single config object case (evals)
-    if hasattr(self.all_params, "tasks") and not isinstance(
-        self.all_params, (list, tuple)
-    ):
-        return 2  # hard coding for evals
-
-    # Handle list/iterable case (benchmarks)
-    if isinstance(self.all_params, (list, tuple)):
-        return next(
-            (
-                getattr(param, "num_eval_runs", 2)
-                for param in self.all_params
-                if hasattr(param, "num_eval_runs")
-            ),
-            2,
-        )
-
-    return 2
 
 
 # =============================================================================
